Add StackExchange.Utils.Configuration
This commit adds a helper library that performs substitution and prefixing for `IConfiguration`-based configuration sources. It allows a value in the configuration tree to reference other values within the configuration system using a placeholder syntax `${key}` or `${nested:key}`. In addition, prefixing allows a "namespace" prefix to be applied to subset of configuration, making it possible to segment configuration into logical areas.

This is particularly useful for storing secrets in a different, secure location but in a way that it makes it easy to compose configuration like connection strings with dealing with it inside the application. E.g. consider the following JSON:

**appsettings.json**
```json
{
    "ConnectionStrings": {
        "Database": "Server=srv01;Database=db01;User ID=${secrets:Database:UserId};Password=${secrets:Database:Password}"
    }
}
```

**secrets.json**
```json
{
    "Database": {
        "UserId": "User123",
        "Password": "Password123!"
    }
}
```

This instructs the substitution provider to lookup the keys `secrets:Database:UserId` and `secrets:Database:Password` from configuration again and replaces them in the value returned for `ConnectionStrings:Database`.

To support this a `ConfigurationBuilder` is configured as follows:

```c#
var configuration = new ConfigurationBuilder()
    .WithPrefix(
        "secrets",
        // everything in this configuration builder will be prefixed with "secrets:"
        c => c.AddJsonFile("secrets.json")
    )
    .WithSubstitution(
        // values in this configuration builder will have substitutions
        // replaced prior to being returned to the caller
        c => c.AddJsonFile("appsettings.json")
    )
    .Build();
```

Here we're loading a JSON file called `secrets.json` (it could equally be any source supported by the `IConfiguration` system) and prefixing it with `secrets:`. Then, we load `appsettings.json` with support for substitutions. If a caller asks the `IConfiguration` that is produced for a connection string:

```c#
var connectionString = configuration.GetConnectionString("Database");
```

That value will be returned as `Server=srv01;Database=db01;User ID=User123;Password=Password123!`.

- Build scripts updated to build all projects
- Builds scripts updated to support Linux/MacOS
- `.gitignore` tweaks for Rider
- Move to `LangVersion = latest`
- `version.json` in each packages directories
- Use `Microsoft.NETFramework.ReferenceAssemblies` for net462 builds on Linux