GitHub Actions support
This PR adds support for using GitHub Actions to build the project. It also adds `Build.csproj` which is a `Microsoft.Build.Traversal` used for building / testing / packaging everything and updates `global.json` to use .NET Core SDK 3.1.302.

There are two GH workflows - one for building and testing (run on every PR to master) and one for building, testing and packaging (used for publishing to MyGet).