Use HttpCompletionOption.ResponseHeadersRead to reduce allocations (#18)
`HttpClient.SendAsync` by default uses `HttpCompletionOption.ResponseContentRead` which buffers the entire response into a `MemoryStream` (creating a copy) before returning. 

`HttpCompletionOption.ResponseHeadersRead` returns control after the headers are read and then if the stream is needed, whenever `HttpResponseMessage.Content.ReadAsStreamAsync()` is called, the network stream is returned instead of the in-memory buffered copy. 

I did some benchmarks before and after using `HttpCompletionOption.ResponseContentRead` in the different `Expect` extension methods. These were the results:

Before:

![image](https://user-images.githubusercontent.com/6766854/88821617-479e9c80-d199-11ea-9639-97ac6905398b.png)

After:

![image](https://user-images.githubusercontent.com/6766854/88822010-b845b900-d199-11ea-969e-7c443c9d1e52.png)


Note: All requests were done to this url: "https://jsonplaceholder.typicode.com/comments"

The biggest impact was on `ExpectHttpSuccess` and on `ExpectJson`, so this PR proposes using `HttpCompletionOption.ResponseContentRead` there. I wasn't able to test `ExpectProtobuf`, but it's possible that it could benefit of it as well.