Configuration: Fix `GetChildKeys` to work correctly in prefixed configuration provider (#19)
Prior to this commit `GetChildKeys` was using a naive implementation that did not function correctly when there were multiple `IConfigurationProvider` implementations chained to the prefixed configuration.

This commit updates the method to ensure that it
 - returns the prefix and earlier keys if the parent path is null
 - returns just the earlier keys if the parent path is not null and does not have the prefix
 - correctly passes the parent path *without* the prefix to the child configuration providers when the parent path is prefixed

It also adds a test to validate that the fixes are sane.

Finally, in a fun episode of Yak Shaving, this tweaks `global.json` and `Directory.Build.props` to ensure that the project can build - there's an issue with upgrading the latest stable VS to 16.8 that inadvertently (and incorrectly) nukes SDK directories and our global.json was preventing the use of the 5.0 SDK for build purposes. And then Nerdbank got angry so needed an upgrade too. Yay!