Fix #23 by using `ThreadLocal` instead of `AsyncLocal` (#25)
If the `SubstitutionHelper`was accessed on a particular async context then the `ICycleDetector` would be propagated to any subsequently async contexts initiated from that first one. `CycleDetector` uses a `HashSet` which would then be accessed concurrently and explode. This was particularly noticeable in Kestrel where the thread that initiated the `WebHost` would access a substituted configuration value, initialise the `ICycleDetector` and then each request thread would then inherit that instance.

Originally the intent was that there'd be only a single `CycleDetector` per thread so this fixes the issue by using a `ThreadLocal` instead. This ensures the cycle detection continues to work but without the concurrency issues that arise from the `AsyncLocal` being inadvertently shared across async contexts.

This commit also does some rudimentary clean-up (`new()` syntax).